requirements.txt
pandas==1.0.1
catboost==0.24.4
streamlit==0.76.0
numpy==1.18.1
Pillow==8.1.1
